5.2- Extracting ChatMessages Component - #30
