+++ Build-AI-Powered-Apps/4-Project-Building-a-Chatbot/app/packages/client/src/components/chat/ChatMessages.tsx
@@ -1,0 +1,51 @@
+import { useEffect, useRef } from "react";
+import ReactMarkdown from "react-markdown";
+
+export type Message = {
+  content: string;
+  role: "user" | "bot";
+};
+
+type ChatMessagesProps = {
+  messages: Message[];
+};
+
+const ChatMessages = ({ messages }: ChatMessagesProps) => {
+  const lastMessageRef = useRef<HTMLDivElement | null>(null);
+
+  useEffect(() => {
+    lastMessageRef.current?.scrollIntoView({ behavior: "smooth" });
+  }, [messages]);
+
+  const onCopyMessage = (e: React.ClipboardEvent<HTMLParagraphElement>) => {
+    const selection = window.getSelection()?.toString().trim();
+    if (selection) {
+      e.preventDefault();
+      e.clipboardData.setData("text/plain", selection);
+    }
+  };
+
+  return (
+    <div className="flex flex-col gap-3">
+      {messages.map((message, index) => (
+        <div
+          key={index}
+          ref={index === messages.length - 1 ? lastMessageRef : null}
+          onCopy={onCopyMessage}
+          className={`
+            px-3 py-1 rounded-xl max-w-[80%]
+            ${
+              message.role === "user"
+                ? "bg-blue-600 text-white self-end"
+                : "bg-gray-100 text-black self-start"
+            }
+          `}
+        >
+          <ReactMarkdown>{message.content}</ReactMarkdown>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default ChatMessages;
