+++ Build-AI-Powered-Apps/4-Project-Building-a-Chatbot/app/packages/server/services/chat.service.ts
@@ -24,6 +24,7 @@ export const chatService = {
       previous_response_id:
         conversationRepository.getLastResponseId(conversationId),
     });
+    console.log(response);
 
     conversationRepository.setLastResponseId(conversationId, response.id);
     return {
